Turn Metrics Configuration's States field into a pointer (#60)
Signed-off-by: Pulak Kanti Bhowmick <pulak@appscode.com>

<b>Why this PR:</b>
Previously State field was a non pointer type. Though it was a struct type optional field but it had some fields in it's struct those are required. So when MetricsConfiguration resource was tried to deployed using client from go code without 'States' field, an error was sent for those missing required field.

Turning the States field into a pointer type, solved the above issue.